+++ CHANGELOG.md
@@ -1,3 +1,42 @@
+# (Thu Jul 04 2024)
+
+#### 🚀 Enhancement
+
+- `@grafana/create-plugin@4.15.0`
+  - Create Plugin: Dynamic webpack publicPath [#966](https://github.com/grafana/plugin-tools/pull/966) ([@jackw](https://github.com/jackw))
+
+#### Authors: 1
+
+- Jack Westbrook ([@jackw](https://github.com/jackw))
+
+---
+
+# (Thu Jul 04 2024)
+
+#### 🐛 Bug Fix
+
+- `@grafana/create-plugin@4.14.1`
+  - Create Plugin: Fix backend dev env for arm64 hosts [#990](https://github.com/grafana/plugin-tools/pull/990) ([@jackw](https://github.com/jackw))
+
+#### Authors: 1
+
+- Jack Westbrook ([@jackw](https://github.com/jackw))
+
+---
+
+# (Mon Jul 01 2024)
+
+#### 🚀 Enhancement
+
+- `@grafana/create-plugin@4.14.0`
+  - Create-Plugin: Add version comment to bundled Javascript [#970](https://github.com/grafana/plugin-tools/pull/970) ([@jackw](https://github.com/jackw))
+
+#### Authors: 1
+
+- Jack Westbrook ([@jackw](https://github.com/jackw))
+
+---
+
 # (Mon Jun 24 2024)
 
 #### 🚀 Enhancement