+++ packages/create-plugin/CHANGELOG.md
@@ -1,3 +1,39 @@
+# v4.15.0 (Thu Jul 04 2024)
+
+#### 🚀 Enhancement
+
+- Create Plugin: Dynamic webpack publicPath [#966](https://github.com/grafana/plugin-tools/pull/966) ([@jackw](https://github.com/jackw))
+
+#### Authors: 1
+
+- Jack Westbrook ([@jackw](https://github.com/jackw))
+
+---
+
+# v4.14.1 (Thu Jul 04 2024)
+
+#### 🐛 Bug Fix
+
+- Create Plugin: Fix backend dev env for arm64 hosts [#990](https://github.com/grafana/plugin-tools/pull/990) ([@jackw](https://github.com/jackw))
+
+#### Authors: 1
+
+- Jack Westbrook ([@jackw](https://github.com/jackw))
+
+---
+
+# v4.14.0 (Mon Jul 01 2024)
+
+#### 🚀 Enhancement
+
+- Create-Plugin: Add version comment to bundled Javascript [#970](https://github.com/grafana/plugin-tools/pull/970) ([@jackw](https://github.com/jackw))
+
+#### Authors: 1
+
+- Jack Westbrook ([@jackw](https://github.com/jackw))
+
+---
+
 # v4.12.0 (Tue Jun 11 2024)
 
 #### 🚀 Enhancement